+++ LetsBank.Core/Entities/TransactionRecord.cs
@@ -11,12 +11,12 @@ namespace LetsBank.Core.Entities
 		public Guid AccountId { get; set; }
 
 		[Required]
-		public int Amount { get; set; }
+		public decimal Amount { get; set; }
 
 		public TransactionType Type { get; set; } = TransactionType.Deposit;
 
 		[Required]
-		public int FinalBalance { get; set; }
+		public decimal FinalBalance { get; set; }
 
 		public DateTime Date { get; set; } = DateTime.Now;
 	}
